Add files via upload
This Power BI dashboard was developed as part of a data-driven initiative to identify critical healthcare challenges in the Umthatha region and guide the development of a community health intervention app.

The dashboard visualizes key metrics around disease prevalence, treatment access, health program awareness, and geographic barriers to healthcare delivery. It supports decision-making for building a mobile application that addresses the real-world needs of underserved populations.